refactor(webpack_config/client): move client's webpack config to own folder
